move types up a module
